requirements.txt
# Discord
discord.py==2.3.2
PyNaCl==1.5.0

# API Clients
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
apify-client==1.6.2

# Web Scraping
aiohttp==3.9.5
beautifulsoup4==4.12.2
lxml==4.9.3

# Media Processing
yt-dlp==2025.4.30
ffmpeg-python==0.2.0

# Browser Automation (for cookie extraction)
selenium==4.15.2
browser-cookie3==0.19.1

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
psutil>=5.8.0
prometheus-client==0.22.0
cryptography==42.0.5

# Async file operations
aiofiles==23.2.1

# Logging
colorlog==6.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring has been moved to utilities section

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
typing-extensions==4.9.0

# Optional performance improvements
aiodns==3.1.1
# cchardet==2.1.7  # Removed due to Python 3.11 compatibility issues
charset-normalizer==3.4.2  # Alternative to cchardet
brotli==1.1.0

# Flask for stream service
flask==3.0.0
gunicorn==21.2.0